backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow>=9.0.0
numpy>=1.21.0
roboflow>=1.1.0
gunicorn==21.2.0
